log file merged
