--- src/main/java/com/hulzzuk/log/model/service/LogReviewService.java
@@ -1,9 +1,0 @@
-package com.hulzzuk.log.model.service;
-
-import com.hulzzuk.log.model.vo.LogReviewVO;
-
-public interface LogReviewService {
-	int insertLogReview(LogReviewVO review);
-	
-	
-}
